+++ tests/Dropdown/ImmutableTest.php
@@ -39,6 +39,7 @@ final class ImmutableTest extends TestCase
         $this->assertNotSame($dropdown, $dropdown->itemsContainerTag(''));
         $this->assertNotSame($dropdown, $dropdown->splitButtonAttributes([]));
         $this->assertNotSame($dropdown, $dropdown->splitButtonClass(''));
+        $this->assertNotSame($dropdown, $dropdown->splitButtonSpanAttributes([]));
         $this->assertNotSame($dropdown, $dropdown->splitButtonSpanClass(''));
         $this->assertNotSame($dropdown, $dropdown->toggleAttributes([]));
         $this->assertNotSame($dropdown, $dropdown->toggleClass(''));